Refactor Endpoints, Schema
